create gitignore file
