+++ lib/getallurl.rb
